+++ src/components/com_medium/domains/entities/component.php
@@ -47,7 +47,7 @@ class ComMediumDomainEntityComponent extends ComComponentsDomainEntityComponent
                 'scopeable' => array(
                     'class' => 'ComMediumDomainEntityMedium',
                     'type' => 'post', ),
-                'hashtagable' => array(
+                'hashtaggable' => array(
                     'class' => 'ComMediumDomainEntityMedium',
                     'type' => 'post',
                 ),