+++ CHANGELOG.md
@@ -2,8 +2,11 @@ To view the detailed commits log go to https://github.com/anahitasocial/anahita/
 
 Anahita 4.4.1
 =============================
+1. changed: renamed hashtagable to hashtaggable everywhere
+1. changed: the base class tag node identifier is now included in the type field of hashtags, locations, or any tags extending the tag node.
+1. changed: tag node (hashtag, location) json response now contains list of taggables with pagination 
 1. added: fake total number for the stories to save on the query operation.
-2. added: owner field to the story json object
+1. added: owner field to the story json object
 
 Anahita 4.4.0
 =============================